fix ctrlContacts

+++ controllers/contacts.js
@@ -6,21 +6,25 @@ const getAll = async (req, res) => {
   const { _id: owner } = req.user;
   const { page = 1, limit = 20, favorite = null } = req.query;
   const skip = (page - 1) * limit;
-  const result = await Contact.find({ owner }, null, {
+
+  const searchParams = favorite ? { owner, favorite } : { owner };
+  const result = await Contact.find(searchParams, "-_id -__v", {
     skip,
     limit,
-  }).populate("owner");
-  if (favorite) {
-    result.map((item) => (item.favorite = favorite));
-  }
+  }).populate("owner", "email subscription -_id");
+
   res.status(200).json(result);
 };
 
 const getById = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
   const result = await Contact.findById(contactId);
   if (!result) {
     throw HttpError(404, "Not found");
+  }
+  if (owner !== result._id) {
+    throw HttpError(403, "Forbidden, this contact does not belong to you ");
   }
   res.status(200).json(result);
 };
@@ -33,7 +37,12 @@ const add = async (req, res) => {
 };
 
 const deleteById = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
+  const itemToDell = await Contact.findById(contactId);
+  if (itemToDell._id !== owner) {
+    throw HttpError(403, "Forbidden, this contact does not belong to you");
+  }
   const result = await Contact.findByIdAndDelete(contactId);
   if (!result) {
     throw HttpError(404, "Not found");
@@ -42,7 +51,12 @@ const deleteById = async (req, res) => {
 };
 
 const updateById = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
+  const itemToUpd = await Contact.findById(contactId);
+  if (itemToUpd._id !== owner) {
+    throw HttpError(403, "Forbidden, this contact does not belong to you");
+  }
   const result = await Contact.findByIdAndUpdate(contactId, req.body, {
     new: true,
   });
@@ -54,7 +68,12 @@ const updateById = async (req, res) => {
 };
 
 const updateStatusContact = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
+  const itemToUpd = await Contact.findById(contactId);
+  if (itemToUpd._id !== owner) {
+    throw HttpError(403, "Forbidden, this contact does not belong to you");
+  }
   const result = await Contact.findByIdAndUpdate(contactId, req.body, {
     new: true,
   });
